+++ backend/modules/blog/actions/add.php
@@ -17,7 +17,7 @@ use Symfony\Component\Filesystem\Exception\IOException;
  * @author Dave Lens <dave.lens@netlash.com>
  * @author Tijs Verkoyen <tijs@sumocoders.be>
  * @author Matthias Mullie <forkcms@mullie.eu>
- * @author Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author Jelmer Snoeck <jelmer@siphoc.com>
  */
 class BackendBlogAdd extends BackendBaseActionAdd
 {