+++ backend/modules/faq/ajax/sequence_questions.php
@@ -5,7 +5,7 @@
  *
  * @author Lester Lievens <lester.lievens@netlash.com>
  * @author Annelies Van Extergem <annelies.vanextergem@netlash.com>
- * @author Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author Jelmer Snoeck <jelmer@siphoc.com>
  */
 class BackendFaqAjaxSequenceQuestions extends BackendBaseAJAXAction
 {