+++ backend/modules/pages/actions/add.php
@@ -13,7 +13,7 @@
  * @author Matthias Mullie <forkcms@mullie.eu>
  * @author Tijs Verkoyen <tijs@sumocoders.be>
  * @author Davy Hellemans <davy.hellemans@netlash.com>
- * @author Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author Jelmer Snoeck <jelmer@siphoc.com>
  */
 class BackendPagesAdd extends BackendBaseActionAdd
 {