+++ backend/modules/search/engine/model.php
@@ -16,7 +16,7 @@ use Symfony\Component\Filesystem\Exception\IOException;
  * In this file we store all generic functions that we will be using in the search module
  *
  * @author Matthias Mullie <forkcms@mullie.eu>
- * @author Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author Jelmer Snoeck <jelmer@siphoc.com>
  */
 class BackendSearchModel
 {