+++ frontend/modules/faq/js/faq.js
@@ -2,7 +2,7 @@
  * Interaction for the faq module
  *
  * @author	Annelies Van Extergem <annelies@netlash.com>
- * @author	Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author	Jelmer Snoeck <jelmer@siphoc.com>
  * @author	Thomas Deceuninck <thomas@fronto.be>
  */
 jsFrontend.faq =