+++ frontend/modules/location/engine/model.php
@@ -11,7 +11,7 @@
  * In this file we store all generic functions that we will be using in the location module
  *
  * @author Matthias Mullie <forkcms@mullie.eu>
- * @author Jelmer Snoeck <jelmer.snoeck@netlash.com>
+ * @author Jelmer Snoeck <jelmer@siphoc.com>
  */
 class FrontendLocationModel
 {