+++ install/engine/step_7.php
@@ -107,40 +107,6 @@ class InstallerStep7 extends InstallerStep
 	}
 
 	/**
-	 * Writes a config file to app/config/parameters.yml.
-	 */
-	private function createYAMLConfig()
-	{
-		// these variables should be parsed inside the config file(s).
-		$variables = $this->getConfigurationVariables();
-		$fs = new Filesystem();
-
-		// map the config templates to their destination filename
-		$yamlFiles = array(
-			PATH_LIBRARY . '/parameters.base.yml' => PATH_WWW . '/app/config/parameters.yml',
-		);
-
-		foreach($yamlFiles as $sourceFilename => $destinationFilename)
-		{
-			$yamlContent = file_get_contents($sourceFilename);
-			$yamlContent = str_replace(
-				array_keys($variables),
-				array_values($variables),
-				$yamlContent
-			);
-
-			// write app/config/parameters.yml
-			$fs->dumpFile($destinationFilename, $yamlContent);
-		}
-
-		// we already went through AppKernel once. Now that our config files are
-		// available, we can register our container configuration.
-		$this->getKernel()->registerContainerConfiguration(
-			$this->getKernel()->getContainerLoader($this->getKernel()->getContainer())
-		);
-	}
-
-	/**
 	 * Create locale cache files
 	 */
 	private function createLocaleFiles()
@@ -216,9 +182,6 @@ class InstallerStep7 extends InstallerStep
 		// delete cached data
 		$this->deleteCachedData();
 
-		// create configuration files
-		$this->createYAMLConfig();
-
 		// define paths
 		$this->definePaths();
 
@@ -240,31 +203,6 @@ class InstallerStep7 extends InstallerStep
 
 		// clear session
 		SpoonSession::destroy();
-	}
-
-	/**
-	 * @return array A list of variables that should be parsed into the configuration file(s).
-	 */
-	protected function getConfigurationVariables()
-	{
-		return array(
-			'<debug-mode>'				=> SpoonSession::get('debug_mode') ? 'true' : 'false',
-			'<debug-email>'				=> SpoonSession::get('different_debug_email') ? SpoonSession::get('debug_email') : SpoonSession::get('email'),
-			'<database-name>'			=> SpoonSession::get('db_database'),
-			'<database-host>'			=> addslashes(SpoonSession::get('db_hostname')),
-			'<database-user>'			=> addslashes(SpoonSession::get('db_username')),
-			'<database-password>'		=> addslashes(SpoonSession::get('db_password')),
-			'<database-port>'			=> (SpoonSession::exists('db_port') && SpoonSession::get('db_port') != '') ? addslashes(SpoonSession::get('db_port')) : 3306,
-			'<site-protocol>'			=> isset($_SERVER['SERVER_PROTOCOL']) ? (strpos(strtolower($_SERVER['SERVER_PROTOCOL']), 'https') === false ? 'http' : 'https') : 'http',
-			'<site-domain>'				=> (isset($_SERVER['HTTP_HOST'])) ? $_SERVER['HTTP_HOST'] : 'fork.local',
-			'<site-default-title>'		=> 'Fork CMS',
-			'<site-multilanguage>'		=> SpoonSession::get('multiple_languages') ? 'true' : 'false',
-			'<site-default-language>'	=> SpoonSession::get('default_language'),
-			'<path-www>'				=> PATH_WWW,
-			'<path-library>'			=> PATH_LIBRARY,
-			'<action-group-tag>'		=> '@actiongroup',
-			'<action-rights-level>'		=> 7
-		);
 	}
 
 	/**